Use store to reactively show room data in RoomsComponent.
The roomsComponent on init does two things:

First, calls store.select('rooms') and stores the result value in
rooms$. That value is an observable of the rooms array in our
store (the result of a call of .asObservable() on an istance of
BehaviorSubject);

Second, it subscribes to the getRoom$ (using the roomService).

As far as I understand, now the roomsComponent is ``listening'' to any
change on the rooms array in the store.

In the rooms component template we use the async pipe to reactively
show the values in the rooms array.