Add filter property in the store and keep it updated.
Add filter property in store which holds the current values of the
user inputs in the room-filter component.

At the moment the logic for updating the view has not changed (we are
not using the store).

The next step will be basing that logic on the current state of the
filter property in the store. The rooms component is already listening
to any change of it.