Fix checkboxes.
Fix checkboxes that were not working... the router link on the parent
element was preventing the click on the checkbox to have any
effect. I've solved the problem by adding, first, a listener on the
click event (on the checkbox) which calls the stopProgation function
and, second, a listener on the change event (on the checkbox) that
handles the changes through the Angular Material api.