Bug fix.
When using the filter, the room component was receiving the filtered
rooms arrays more than once.

The problem was that when a certain property in the store was
updated (using store.set), all observablers of any store property were
emitting a new value.

Fix: using distinctUntilChanged in the store.select method.